Support for comments on views for Delta Lake
